- Confirmations de création et d'effacement - Mise en page dialogue de création - Optimisations

+++ frontend/scripts/index.js
@@ -3,11 +3,11 @@
 const API_URL = 'http://localhost:3000/articles';
 
 window.onload = function () {
-
+  // Load all citations when DOM document ready
   loadCitations();
 };
 
-// Load citations from remote server
+// Load citations from the remote server
 function loadCitations() {
 
   let xhr = new XMLHttpRequest();
@@ -49,12 +49,12 @@ function deleteCitation(id) {
 function showCitations(citations) {
 
   var container = document.getElementById('citations');
-  var model = document.getElementById('model');
+  var model = document.getElementById('template');
 
   for (let citation of citations) {
 
     var source = `${citation.author}, <i>${citation.chapter}, ${citation.episode}</i>`;
-    var thumb = container.appendChild(model.cloneNode(true));
+    var thumb = model.cloneNode(true);
 
     thumb.getElementsByClassName('source')[0].innerHTML = source;
     thumb.getElementsByTagName('blockquote')[0].textContent = citation.content;
@@ -64,31 +64,43 @@ function showCitations(citations) {
     thumb.getElementsByClassName('actionDelete')[0].onclick = actionDelete;
 
     thumb.id = citation.id;
+
+    container.appendChild(thumb);
   }
 }
 
 // Display citation details modal dialog box
 function actionDetails(e) {
 
-  var thumb = getParentThumb(e.currentTarget.parentNode); 
-  var modal = document.getElementById('seeMoreModal');
+  var thumb = getParentThumb(e.currentTarget.parentNode);
 
-  var content = thumb.getElementsByTagName('blockquote')[0].textContent;
-  var source = thumb.getElementsByClassName('source')[0].innerHTML;
+  if (thumb) {
+    var modal = document.getElementById('seeMoreModal');
 
-  modal.getElementsByTagName('blockquote')[0].textContent = content;
-  modal.getElementsByClassName('source')[0].innerHTML = source;
+    var content = thumb.getElementsByTagName('blockquote')[0].textContent;
+    var source = thumb.getElementsByClassName('source')[0].innerHTML;
+
+    modal.getElementsByTagName('blockquote')[0].textContent = content;
+    modal.getElementsByClassName('source')[0].innerHTML = source;
+  }
 }
 
 // Delete citation
 function actionDelete(e) {
-  var thumb = getParentThumb(e.currentTarget.parentNode); 
-  deleteCitation(thumb.id);
+
+  e.preventDefault();
+
+  var thumb = getParentThumb(e.currentTarget);
+
+  if (thumb && confirm('Etes-vous sûr(e) de vouloir effacer cette citation ?')) {
+    deleteCitation(thumb.id);
+  }
 }
 
+// Search parent frame givent 
 function getParentThumb(element) {
-  
-  while (element && !element.classList.contains('thumb')) {
+
+  while (element && element.nodeName !== 'ARTICLE') {
     element = element.parentNode;
   }
